Move open base message type enums to the common package
